+++ packages/f-form-field/CHANGELOG.md
@@ -3,6 +3,26 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.8.0
+------------------------------
+*October 23, 2020*
+
+### Added 
+- data-test IDs to FormField Component 
+- Test for FormField 
+- FormField Component-Object 
+
+
+Latest (to be added to next release)
+------------------------------
+*October 26, 2020*
+
+### Added
+- Stylelint added to lint styling on build.
+
+### Changed
+- 'jet' theme instead of 'je'
+
 
 v0.7.1
 ------------------------------