+++ packages/f-mega-modal/CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+Latest (to be added to next release)
+------------------------------
+*October 26, 2020*
+
+### Added
+- Stylelint added to lint styling on build.
+
+### Changed
+- 'jet' theme instead of 'je'
+
+
 v0.2.0
 ------------------------------
 *August 6, 2020*